Updated atoi / strlen

+++ 01 libft/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: cwan <marvin@42.fr>                        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/14 16:26:12 by cwan              #+#    #+#             */
-/*   Updated: 2023/08/16 13:43:41 by cwan             ###   ########.fr       */
+/*   Updated: 2023/09/08 10:33:16 by cwan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(char *str)
+#include <stddef.h>
+
+size_t	ft_strlen(char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
